FIR/UAST: split KotlinUIdentifier into a separate file
